get zip to generate async
